refactor(session): extract useAnalysisRunners from DesignSession
The four background analyses that shadow the live solve — freq sweep,
segments-per-wire convergence, far-field norm check and the NEC rp_card
pattern — move whole into components/session/useAnalysisRunners.ts with their
debounce effects, timer/abort refs, streaming runners and the
CONVERGE_N_VALUES ladder (now exported, since the gear menu and the Smith
overlay still render it).

All four literal dep arrays are byte-identical moves. Every dep-array member
arrives as a plain per-render value and buildRequest / solveWithheld as plain
per-render functions — no useCallback, because memoizing either would change
which closure a pending 500 ms timeout fires. The hook is called at the
debounce effects' old position, so all four keep their global effect order
behind the solve and preview effects.
